requirements.txt
pandas
psycopg2-binary
pytest
great_expectations
streamlit
